Update export-answers for LTI and new folder structure
This echoes the changes in auto-importer.ts to default to keying the
answers from the LTI info if possible, and to upload the parquet files
to the new folder structure.

All the new logic is imported from the shared ts library.

This also changes the script so that the bucket and source names are
brought in as arguments when calling the script.

+++ scripts/export-answers.js
@@ -1,24 +1,27 @@
 #!/usr/bin/env node
-
-const BUCKET = "concordqa-report-data"
-const SOURCE = "authoring.staging.concord.org"
-const DIRECTORY = "partitioned-answers"
-const BATCH_SIZE = 100;
-
-var AWS = require('aws-sdk');
+const AWS = require('aws-sdk');
 const {Firestore} = require('@google-cloud/firestore');
 const fs = require('fs');
 const path = require('path');
-
-const zlib = require('zlib');
 const util = require('util');
+
+const BUCKET = process.argv[2];     // e.g. concordqa-report-data
+const SOURCE =  process.argv[3];    // e.g. authoring.staging.concord.org
+
+if (!BUCKET || !SOURCE) {
+  console.error("Call script with `node export-answers.js [bucket-name] [source-id]`");
+  return;
+}
+
+const DIRECTORY = "partitioned-answers"
+const BATCH_SIZE = 100;
 
 const access = util.promisify(fs.access);
 const unlink = util.promisify(fs.unlink);
 const readFile = util.promisify(fs.readFile);
 
 const parquet = require('parquetjs');
-const { parquetInfo, schema } = require('../functions/src/shared/s3-answers')
+const { parquetInfo, getSyncDocId, schema } = require('../functions/src/shared/s3-answers')
 
 const awsConfig = "./config.json"
 if (!fs.existsSync(awsConfig)) {
@@ -45,31 +48,41 @@ const firestore = new Firestore();
 
 let query = firestore.collection(`sources/${SOURCE}/answers`);
 let masterQuery = query.limit(BATCH_SIZE)
-  .orderBy('run_key').orderBy("id")
+  .orderBy("resource_url")
+  .orderBy("platform_id")
+  .orderBy("resource_link_id")
+  .orderBy("platform_user_id")
+  .orderBy("run_key")
+  .orderBy("id")
 
 let count = 0;
 let lastDoc = null;
 let batchCount = 0;
 let batchIndex = 0;
-let lastRunKey = undefined;
+let lastUserRunKey = undefined;   // id for the a user's whole answer set, whether a run key or an hashed LTI combo
 const answerHash = {};
 const uploadPromises = {};
 
 // set to true once all uploads have been created and the system is waiting for the uploads to finish
 let waitingForUploadsToFinish = false;
 
-function uploadAnswers(runKey) {
-  if (!answerHash[runKey] || (answerHash[runKey].length === 0)) {
+function uploadAnswers(userRunKey) {
+  if (!answerHash[userRunKey] || (answerHash[userRunKey].length === 0)) {
     return;
   }
 
-  const answers = answerHash[runKey];
+  const answers = answerHash[userRunKey];
   const answer = answers[0];
-  const {filename, key} = parquetInfo(answer, DIRECTORY);
+  const info = parquetInfo(DIRECTORY, answer);
+  if (!info) {
+    return;
+  }
+  const {filename, key} = info;
+
   const tmpFilePath = path.join(outputPath, filename);
 
-  if (uploadPromises[runKey]) {
-    console.error("Already uploading", runKey)
+  if (uploadPromises[userRunKey]) {
+    console.error("Already uploading", userRunKey)
     return;
   }
 
@@ -77,13 +90,19 @@ function uploadAnswers(runKey) {
     return access(tmpFilePath).then(() => unlink(tmpFilePath)).catch(() => undefined);
   }
 
-  uploadPromises[runKey] = new Promise(async (resolve) => {
+  uploadPromises[userRunKey] = new Promise(async (resolve) => {
     try {
       // parquetjs can't write to buffers
       await deleteTmpFile();
       var writer = await parquet.ParquetWriter.openFile(schema, tmpFilePath);
       for (let i = 0; i < answers.length; i++) {
+        // clean up answer objects for parquet
         answers[i].answer = JSON.stringify(answers[i].answer)
+        delete answers[i].report_state;
+        if (typeof answers[i].version === "number") {
+          answers[i].version = "" + answers[i].version;
+        }
+
         await writer.appendRow(answers[i]);
       }
       await writer.close();
@@ -99,21 +118,21 @@ function uploadAnswers(runKey) {
         ContentType: 'application/octet-stream'
       }, function (err) {
         if (err) {
-          console.error(`${runKey}: ${err.toString()}`)
+          console.error(`${userRunKey}: ${err.toString()}`)
         }
       }).promise();
     } catch (err) {
       console.error(err);
     } finally {
       await deleteTmpFile();
-      delete answerHash[runKey];
+      delete answerHash[userRunKey];
     }
 
     // while running keep the upload promises object as small as possible by removing them from the hash
     // this is set to true once all the uploads have been created and the system is waiting for them to complete
     // via Promise.allSettled(...)
     if (!waitingForUploadsToFinish) {
-      delete uploadPromises[runKey];
+      delete uploadPromises[userRunKey];
     }
 
     // always resolve
@@ -124,16 +143,16 @@ function uploadAnswers(runKey) {
 function countBatch(query) {
   query.stream().on('data', (documentSnapshot) => {
     const answer = documentSnapshot.data();
-    answerHash[answer.run_key] = answerHash[answer.run_key] || [];
-    answerHash[answer.run_key].push(answer);
-    if (lastRunKey && (answer.run_key !== lastRunKey)) {
-      uploadAnswers(lastRunKey)
+    const userRunKey = getSyncDocId(answer);
+    answerHash[userRunKey] = answerHash[userRunKey] || [];
+    answerHash[userRunKey].push(answer);
+    if (lastUserRunKey && (userRunKey !== lastUserRunKey)) {
+      uploadAnswers(lastUserRunKey)
     }
-    lastRunKey = answer.run_key;
+    lastUserRunKey = userRunKey;
     lastDoc = documentSnapshot;
     ++count;
     ++batchCount;
-  // }).on('end', () => console.log('Done'));
   }).on('end', onEnd);
 }
 
@@ -142,7 +161,7 @@ function finish() {
   // flag that we are waiting for the uploads to finish - this changes the uploader to not remove upload promises from the hash it uses
   // to track them so that we can then use Promise.allSettled to wait for them to finish
   waitingForUploadsToFinish = true;
-  uploadAnswers(lastRunKey);
+  uploadAnswers(lastUserRunKey);
   const promises = Object.values(uploadPromises);
   console.log(`Waiting for ${promises.length} uploads to finish...`);
   Promise.allSettled(promises).then(() => console.log("Done waiting!"))
